+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Recommendation: for ease of reading, use the following order:
 ## Unreleased
 ### Changed
 - Simple Transfer Protocol & Smart Transfer Protocol use `AppendDatasetMetadataBatchUseCase`
+- SQLite: protection against database locking, in case of parallel execution of `kamu` commands.
+  - Based on `journal_mode=WAL`
 
 ## [0.226.3] - 2025-02-27
 ### Changed